SubregionRequirement → module
